+++ utils/image_metrics_utils.go
@@ -2,6 +2,6 @@ package utils
 
 import "image"
 
-func GetSize(sourceImage image.Image) (width int64, height int64) {
-	return int64(sourceImage.Bounds().Size().X), int64(sourceImage.Bounds().Size().Y)
+func GetSize(sourceImage image.Image) (width uint32, height uint32) {
+	return uint32(sourceImage.Bounds().Size().X), uint32(sourceImage.Bounds().Size().Y)
 }
